create test for filter component
