+++ index.html
@@ -16,6 +16,6 @@
         <li><input id = "choices-ui" placeholder="pokes per page"></input></li>
     </ul>
     <ul id='card-gallery'></ul>
-    <script type="module" src="./index.js"></script>
+    <script type="module" src="../src/index.js"></script>
 </body>
 </html>
