+++ test/tests.js
@@ -1,3 +1,4 @@
 import './html-equal.js';
 import './create-list-component.test.js';
 import './paginate.test.js';
+import './filter.test.js';
